root page uses data context
