+++ app/page.js
@@ -1,28 +1,58 @@
-import Home from './Home';
-// import { ENDPOINTS } from '../helpers/constants';
+'use client';
 
-// async function getData() {
-//   const res = await fetch(
-//     'https://pluspool-east-river-data.s3.us-east-2.amazonaws.com/2020-04-30T00%3A00%3A46.330Z.json',
-//     {
-//       method: 'GET',
-//       mode: 'cors',
-//       cache: 'no-cache',
-//       credentials: 'same-origin',
-//       headers: {
-//         'Content-Type': 'application/json',
-//         'Accept-Encoding': 'gzip',
-//       },
-//     }
-//   );
+import React, { useState, useContext } from 'react';
 
-//   return res.json();
-// }
+import Databar from '../components/Databar';
+import DataRangePicker from '../components/DataRangePicker';
+import TitleText from '../components/TitleText';
+import SvgVisualization from '../components/SvgVisualization';
+import Tooltip from '../components/Tooltip';
 
-const Page = async () => {
-  // const { sources, samples, units } = await getData();
+import { useSample } from '../hooks/useSamples';
+import { DataContext } from '../providers/DataProvider';
 
-  return <Home />;
+const Index = () => {
+  const { samples, sources, units } = useContext(DataContext);
+
+  const [tooltipOpen, setTooltipOpen] = useState(false);
+  const [tooltipSlug, setTooltipSlug] = useState();
+
+  const [sample, range, timestamp, setTimestamp] = useSample(samples);
+  const [pageState, setPageState] = useState(0);
+
+  const advanceIntro = () => pageState < 1 && setPageState(pageState + 1);
+
+  const openTooltip = (slug) => {
+    setTooltipSlug(slug);
+    setTooltipOpen(true);
+  };
+  const closeTooltip = () => setTooltipOpen(false);
+
+  return (
+    <main className='page' data-template='index' data-page-state={pageState}>
+      <Tooltip
+        open={tooltipOpen}
+        slug={tooltipSlug}
+        sample={sample}
+        closeTooltip={closeTooltip}
+        sources={sources}
+        units={units}
+      />
+      <TitleText
+        timestamp={timestamp}
+        sample={sample}
+        pageState={pageState}
+        onClick={advanceIntro}
+      />
+      <DataRangePicker
+        setTimestamp={setTimestamp}
+        timestamp={timestamp}
+        range={range}
+      />
+      <Databar openTooltip={openTooltip} sample={sample} />
+      <SvgVisualization sample={sample} />
+    </main>
+  );
 };
 
-export default Page;
+export default Index;
